+++ RELEASENOTES.md
@@ -7,3 +7,4 @@
   - Logic error with enemizer and standard should use new enemy logic rules
   - Fixed a bug with the inconsistent treatment of the beemizer setting
   - Fixed an issue with returning Blacksmith in Simple shuffle (when blacksmith is at Link's House)
+  - Fixed an issue with dark sanc spawn at tavern north door (thanks Codemann!)